+++ common/src/main/java/io/github/mortuusars/scholar/mixin/BlockEntityParentMixin.java
@@ -1,6 +1,10 @@
-package io.github.mortuusars.scholar.mixin.chiseled_bookshelf;
+package io.github.mortuusars.scholar.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import com.llamalad7.mixinextras.sugar.Local;
+import io.github.mortuusars.scholar.mixin.chiseled_bookshelf.ChiseledBookShelfBlockEntityMixin;
+import io.github.mortuusars.scholar.mixin.lectern.LecternBlockEntityMixin;
+import net.minecraft.core.HolderLookup;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
@@ -14,13 +18,11 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 
 /**
- * This class is used so child classes {@link ChiseledBookShelfBlockEntityMixin}
+ * This class is used so child classes like {@link ChiseledBookShelfBlockEntityMixin} or {@link LecternBlockEntityMixin}
  * could modify methods in a (hopefully) more compatible with other mods way.
  */
 @Mixin(BlockEntity.class)
 public abstract class BlockEntityParentMixin {
-    @Shadow @Nullable protected Level level;
-
     @ModifyReturnValue(method = "getUpdatePacket", at = @At("RETURN"))
     protected @Nullable Packet<ClientGamePacketListener> onGetUpdatePacket(@Nullable Packet<ClientGamePacketListener> original) {
         return original;